// src/app/models/request.ts
export class Request {
  _id:number;
  uid: String;
  first_name: String;
  last_name: String;
  email: String;
  contact: number;
  yearofgrad: number;
  category: number;
  company?: String;
  pubdetails: String;
  prof: String;
  branch: String;
  status: String;
  level:number;
  reasonofrejection?: String;
  rejectedBy?:String;
}